+++ pilot/pkg/bootstrap/configcontroller.go
@@ -32,6 +32,7 @@ import (
 	"istio.io/istio/pilot/pkg/features"
 	"istio.io/istio/pilot/pkg/leaderelection"
 	"istio.io/istio/pilot/pkg/model"
+	"istio.io/istio/pkg/activenotifier"
 	"istio.io/istio/pkg/adsc"
 	"istio.io/istio/pkg/config/analysis/incluster"
 	"istio.io/istio/pkg/config/schema/collections"
@@ -189,7 +190,7 @@ func (s *Server) initK8SConfigStore(args *PilotArgs) error {
 		}
 	}
 	if features.EnableAmbientStatus {
-		statusWritingEnabled := model.NewActiveNotifier(false)
+		statusWritingEnabled := activenotifier.New(false)
 		args.RegistryOptions.KubeOptions.StatusWritingEnabled = statusWritingEnabled
 		s.addTerminatingStartFunc("ambient status", func(stop <-chan struct{}) error {
 			leaderelection.
