+++ pilot/pkg/serviceregistry/kube/controller/ambient/ambientindex.go
@@ -30,6 +30,7 @@ import (
 	"istio.io/istio/pilot/pkg/features"
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/serviceregistry/kube/controller/ambient/statusqueue"
+	"istio.io/istio/pkg/activenotifier"
 	"istio.io/istio/pkg/cluster"
 	"istio.io/istio/pkg/config/constants"
 	"istio.io/istio/pkg/config/labels"
@@ -120,7 +121,7 @@ type Options struct {
 	XDSUpdater            model.XDSUpdater
 	LookupNetwork         LookupNetwork
 	LookupNetworkGateways LookupNetworkGateways
-	StatusNotifier        *model.ActiveNotifier
+	StatusNotifier        *activenotifier.ActiveNotifier
 }
 
 func New(options Options) Index {
